requirements.txt
pandas
mlflow==2.2.2
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
Flask
Flask-Cors
fastapi
uvicorn
zenml[server]
tensorflow
keras
catboost
lightgbm
markupsafe
xgboost
streamlit
dagshub

-e .